requirements.txt
python-dotenv==1.0.1
feedparser==6.0.11
requests==2.32.3
tqdm==4.66.4
openai==1.45.0
